requirements.txt
anthropic
aiohttp
python-dotenv
pydantic
uuid
pathlib
bs4
apscheduler
asyncio